[tests] Move CreateRandomTx to testutil
